requirements.txt
python-telegram-bot
python-libtorrent
requests
python-dotenv
qrcode
